feat(config): use npm config object, deprecate czConfig
Npm is generous enough to give us a config object. We should use it. This change is not breaking but
sets up 3.0.0 to be breaking by adding a deprecation warning to any repos using the old style of
config. czConfig will be removed in 3.0.0. From now on, commitizen init now uses the new config as
its default. Adds additional tests and docs as well

Closes #30