fix(config): fixes incorrect deprecation notice
In certain scenarios a deprecation warning for czConfig was shown, even if the user had actually
correctly configured commitizen to use the config.commitizen key. This fixes that issue and improves
the warning in the deprecation notice to be more clear that it is specifically the repo (and not the
adapter or the cli) that is having issues.

Closes #50