chore(tests): adds appveyor test config
We need to run our test suite against both windows and linux. In order to accomplish this we need
this config so that when linux tests are done, the windows tests on appveyor are triggered.
Unfortunately since we have to use travis encrypted environment variables, it will be a bit trickier
for people to run the test suite locally since pull request tests will fail.