chore(tests): remove node 0.10 from tests config
Node 4 has an LTS release. Node 0.12 has been out for some time. If we want to offer 0.10 support
then we need to add further dependencies as polyfills. This tradeoff makes it clear that it is time
to remove node 0.10 from our test suite. Everything will still work for now, but from here forward
we'll support only 0.12 and up.

Closes #64