chore(tests): fixes OS-based difference in git log output
Windows git and *nix git output of the log command treats empty lines differently. Windows trims the
spaces while *nix leaves them intact. This means that when comparing output in a test, we need to
take into account whether the output of the os when determining if the test failed or not.