fix(commit): fixes Windows multiline commits
Windows git command line will not handle multiline commit messages passed to it regularly. Instead,

we must add multiple -m arguments, one for each line. I landed this feature in git-gulp previously

but we needed to detect the os and pass an array of lines to git-gulp in order for it to use this

method. This change makes sure that we use the -m method when commiting on windows.