fix(node): remove node 6 and 8 support
This removes node 6 and 8 support. This is a free project and our builds are getting very long so we
have to be a bit careful about what versions we test on. To be clear, this still works on Node 6 and
8 for now but we won't be maintaining that guarentee into the future so please work towards
updating.

BREAKING CHANGE: Removal of two major node versions, 6 and 8 from testing scheme