fix(node): remove node 6 and 8 support (#649)
This removes node 6 and 8 support. This is a free project and our builds are getting very long so we
have to be a bit careful about what versions we test on. To be clear, this still works on Node 6 and
8 for now but we won't be maintaining that guarantee into the future so please work towards
updating. This also removes eslint, likely in favor of prettier at some future date.

BREAKING CHANGE: Removal of two major node versions, 6 and 8 from testing scheme