docs(README.md): fix local husky git hook snippet
The snippet to setup husky to run cz when `git commit` is run assumes the developer has setup the command `git cz` globally and throws an error: `git: 'cz' is not a git command. See 'git --help'.`

The fix references the locally installed cz as the git hook example above it.